feat: refactor schema helper
