+++ src/View/Helper/SchemaHelper.php
@@ -13,6 +13,7 @@
 
 namespace App\View\Helper;
 
+use App\Core\Util\Form;
 use Cake\Core\Configure;
 use Cake\Utility\Hash;
 use Cake\Utility\Inflector;
@@ -23,236 +24,25 @@ use Cake\View\Helper;
  */
 class SchemaHelper extends Helper
 {
-
-    /**
-     * Infer control type from property schema.
-     *
-     * @param mixed $schema Property schema.
-     * @return string
-     */
-    public function controlTypeFromSchema($schema) : string
-    {
-        if (!is_array($schema)) {
-            return 'text';
-        }
-
-        if (!empty($schema['oneOf'])) {
-            foreach ($schema['oneOf'] as $subSchema) {
-                if (!empty($subSchema['type']) && $subSchema['type'] === 'null') {
-                    continue;
-                }
-
-                return $this->controlTypeFromSchema($subSchema);
-            }
-        }
-
-        if (empty($schema['type'])) {
-            return 'text';
-        }
-
-        switch ($schema['type']) {
-            case 'object':
-                return 'json';
-
-            case 'string':
-                if (!empty($schema['format']) && ($schema['format'] === 'date-time')) {
-                    return 'date-time'; // TODO: replace with "datetime".
-                }
-                if (!empty($schema['format']) && $schema['format'] === 'date') {
-                    return 'date';
-                }
-                if (!empty($schema['contentMediaType']) && $schema['contentMediaType'] === 'text/html') {
-                    return 'textarea';
-                }
-                if (!empty($schema['enum']) && is_array($schema['enum'])) {
-                    return 'enum';
-                }
-
-                return 'text';
-
-            case 'number':
-            case 'integer':
-                return 'number';
-
-            case 'boolean':
-            case 'array':
-                return 'checkbox';
-        }
-
-        return 'text';
-    }
-
-    /**
-     * Get controls option by property name.
-     *
-     * @param string $name The field name.
-     * @return array
-     */
-    protected function customControlOptions($name) : array
-    {
-        switch ($name) {
-            case 'lang':
-                $languages = Configure::read('Project.I18n.languages');
-                if (empty($languages)) {
-                    return [
-                        'type' => 'text',
-                    ];
-                }
-                $options = [];
-                foreach ($languages as $key => $description) {
-                    $options[] = ['value' => $key, 'text' => __($description)];
-                }
-
-                return ['type' => 'select'] + compact('options');
-
-            case 'status':
-                return [
-                    'type' => 'radio',
-                    'options' => [
-                        ['value' => 'on', 'text' => __('On')],
-                        ['value' => 'draft', 'text' => __('Draft')],
-                        ['value' => 'off', 'text' => __('Off')],
-                    ],
-                    'templateVars' => [
-                        'containerClass' => 'status',
-                    ],
-                ];
-
-            case 'old_password':
-                return [
-                    'class' => 'password',
-                    'label' => __('Current password'),
-                    'placeholder' => __('current password'),
-                    'autocomplete' => 'current-password',
-                    'type' => 'password',
-                    'default' => '',
-                ];
-
-            case 'password':
-                return [
-                    'class' => 'password',
-                    'placeholder' => __('new password'),
-                    'autocomplete' => 'new-password',
-                    'default' => '',
-                ];
-
-            case 'confirm-password':
-                return [
-                    'label' => __('Retype password'),
-                    'id' => 'confirm_password',
-                    'name' => 'confirm-password',
-                    'class' => 'confirm-password',
-                    'placeholder' => __('confirm password'),
-                    'autocomplete' => 'new-password',
-                    'default' => '',
-                    'type' => 'password',
-                ];
-
-            case 'title':
-                return [
-                    'class' => 'title',
-                    'type' => 'text',
-                ];
-
-            default:
-                return [];
-        }
-    }
-
     /**
      * Get control options for a property schema.
      *
      * @param string $name Property name.
-     * @param string $value Property value.
+     * @param mixed $value Property value.
      * @param array $schema Object schema array.
      * @return array
      */
     public function controlOptions($name, $value, $schema) : array
     {
-        $options = $this->customControlOptions($name);
+        $options = Form::customControlOptions($name);
         if ($options) {
             $options['value'] = $value;
 
             return $options;
         }
+        $type = Form::controlTypeFromSchema($schema);
 
-        $type = $this->controlTypeFromSchema($schema);
-        if ($type === 'json') {
-            return [
-                'type' => 'textarea',
-                'v-jsoneditor' => '',
-                'class' => 'json',
-                'value' => json_encode($value),
-            ];
-        } elseif ($type === 'textarea') {
-            return [
-                'type' => 'textarea',
-                'v-richeditor' => '',
-                'ckconfig' => 'configNormal',
-                'value' => $value,
-            ];
-        } elseif ($type === 'date-time') {
-            return [
-                'type' => 'text',
-                'v-datepicker' => '',
-                'time' => 'true',
-                'value' => $value,
-            ];
-        } elseif ($type === 'date') {
-            return [
-                'type' => 'text',
-                'v-datepicker' => '',
-                'time' => 'false',
-                'value' => $value,
-            ];
-        } elseif ($type === 'checkbox') {
-            if (!empty($schema['oneOf'])) {
-                $options = [];
-                foreach ($schema['oneOf'] as $one) {
-                    if (!empty($one['type']) && ($one['type'] === 'array')) {
-                        $options = array_map(
-                            function ($value) {
-                                return ['value' => $value, 'text' => Inflector::humanize($value)];
-                            },
-                            (array)Hash::extract($one, 'items.enum')
-                        );
-                    }
-                }
-                if (!empty($options)) {
-                    return [
-                        'type' => 'select',
-                        'options' => $options,
-                        'multiple' => 'checkbox',
-                    ];
-                }
-            }
-
-            return [
-                'type' => 'checkbox',
-                'checked' => $value,
-            ];
-        } elseif ($type === 'enum') {
-            if (!empty($schema['oneOf'])) {
-                foreach ($schema['oneOf'] as $one) {
-                    if (!empty($one['enum'])) {
-                        $schema['enum'] = $one['enum'];
-                        array_unshift($schema['enum'], '');
-                    }
-                }
-            }
-
-            return [
-                'type' => 'select',
-                'options' => array_map(
-                    function ($value) {
-                        return ['value' => $value, 'text' => Inflector::humanize($value)];
-                    },
-                    $schema['enum']
-                ),
-            ];
-        }
-
-        return compact('type', 'value');
+        return Form::control($schema, $type, $value);
     }
 
     /**
